+++ src/core/AABB.cpp
@@ -5,15 +5,15 @@ namespace raytracer
 {
 //----------------------------------------------------------------------------------
 AxisAlignedBoundingBox::AxisAlignedBoundingBox()
-    : m_pMin(glm::vec3(std::numeric_limits<double>::max()))
-    , m_pMax(glm::vec3(std::numeric_limits<double>::lowest()))
-    , m_padding(0.0001)
+    : m_pMin(glm::vec3(std::numeric_limits<float>::max()))
+    , m_pMax(glm::vec3(std::numeric_limits<float>::lowest()))
+    , m_padding(0.0001f)
 {}
 
 //----------------------------------------------------------------------------------
-AxisAlignedBoundingBox::AxisAlignedBoundingBox(const glm::dvec3 &p1,
-                                               const glm::dvec3 &p2,
-                                               double padding)
+AxisAlignedBoundingBox::AxisAlignedBoundingBox(const glm::vec3 &p1,
+                                               const glm::vec3 &p2,
+                                               float padding)
     : m_pMin(glm::vec3(std::min(p1.x, p2.x), std::min(p1.y, p2.y), std::min(p1.z, p2.z)))
     , m_pMax(glm::vec3(std::max(p1.x, p2.x), std::max(p1.y, p2.y), std::max(p1.z, p2.z)))
     , m_padding(padding)
@@ -22,7 +22,7 @@ AxisAlignedBoundingBox::AxisAlignedBoundingBox(const glm::dvec3 &p1,
 }
 
 //----------------------------------------------------------------------------------
-glm::dvec3 AxisAlignedBoundingBox::corner(const int c) const
+glm::vec3 AxisAlignedBoundingBox::corner(const int c) const
 {
     if(c >= 8 || c < 0)
     {
@@ -36,23 +36,23 @@ glm::dvec3 AxisAlignedBoundingBox::corner(const int c) const
     auto y = (c & 2) ? m_pMax.y : m_pMin.y;
     auto z = (c & 4) ? m_pMax.z : m_pMin.z;
 
-    return glm::dvec3(x,y,z);
+    return glm::vec3(x,y,z);
 }
 
 //----------------------------------------------------------------------------------
 bool AxisAlignedBoundingBox::intersect(const Ray &ray) const
 {
-    glm::dvec3 invRayDir = glm::dvec3(1.0 / ray.direction().x,
-                                      1.0 / ray.direction().y,
-                                      1.0 / ray.direction().z);
+    glm::vec3 invRayDir = glm::vec3(1.0f / ray.direction().x,
+                                    1.0f / ray.direction().y,
+                                    1.0f / ray.direction().z);
 
     // Absolute distance to lower and upper box coordinates
-    glm::dvec3 tLower = (m_pMin - ray.origin()) * invRayDir;
-    glm::dvec3 tUpper = (m_pMax - ray.origin()) * invRayDir;
+    glm::vec3 tLower = (m_pMin - ray.origin()) * invRayDir;
+    glm::vec3 tUpper = (m_pMax - ray.origin()) * invRayDir;
 
     // The four t-intervals (for x-/y-/z-slabs, and ray p(t))
-    glm::dvec4 tMins = glm::dvec4(glm::min(tLower, tUpper), ray.tMin());
-    glm::dvec4 tMaxes = glm::dvec4(glm::max(tLower, tUpper), ray.tMax());
+    glm::vec4 tMins = glm::vec4(glm::min(tLower, tUpper), ray.tMin());
+    glm::vec4 tMaxes = glm::vec4(glm::max(tLower, tUpper), ray.tMax());
 
     auto tBoxMin = std::max(std::max(std::max(tMins[0], tMins[1]), tMins[2]), tMins[3]);
     auto tBoxMax = std::min(std::min(std::min(tMaxes[0], tMaxes[1]), tMaxes[2]), tMaxes[3]);
@@ -61,17 +61,17 @@ bool AxisAlignedBoundingBox::intersect(const Ray &ray) const
 }
 
 //----------------------------------------------------------------------------------
-double AxisAlignedBoundingBox::surfaceArea() const
+float AxisAlignedBoundingBox::surfaceArea() const
 {
-    glm::dvec3 diagonal = m_pMax - m_pMin;
+    glm::vec3 diagonal = m_pMax - m_pMin;
     auto retVal = 2 * (diagonal.x * diagonal.y + diagonal.x * diagonal.z + diagonal.y * diagonal.z);
     return retVal;
 }
 
 //----------------------------------------------------------------------------------
-double AxisAlignedBoundingBox::volume() const
+float AxisAlignedBoundingBox::volume() const
 {
-    glm::dvec3 diagonal = m_pMax - m_pMin;
+    glm::vec3 diagonal = m_pMax - m_pMin;
     auto retVal = diagonal.x * diagonal.y * diagonal.z;
     return retVal;
 }
@@ -109,12 +109,12 @@ AxisAlignedBoundingBox AxisAlignedBoundingBox::combine(const AxisAlignedBounding
 
 //----------------------------------------------------------------------------------
 AxisAlignedBoundingBox AxisAlignedBoundingBox::combine(const AxisAlignedBoundingBox &box,
-                                                       const glm::dvec3 &point)
+                                                       const glm::vec3 &point)
 {
-    return AxisAlignedBoundingBox(glm::dvec3(std::min(box.m_pMin.x, point.x),
+    return AxisAlignedBoundingBox(glm::vec3(std::min(box.m_pMin.x, point.x),
                                             std::min(box.m_pMin.y, point.y),
                                             std::min(box.m_pMin.z, point.z)),
-                                  glm::dvec3(std::max(box.m_pMax.x, point.x),
+                                  glm::vec3(std::max(box.m_pMax.x, point.x),
                                             std::max(box.m_pMax.y, point.y),
                                             std::max(box.m_pMax.z, point.z)));
 }
@@ -122,7 +122,7 @@ AxisAlignedBoundingBox AxisAlignedBoundingBox::combine(const AxisAlignedBounding
 //----------------------------------------------------------------------------------
 void AxisAlignedBoundingBox::padToMinExtent()
 {
-    auto padding = m_padding * 0.5;
+    auto padding = m_padding * 0.5f;
 
     if( (m_pMax.x - m_pMin.x) < m_padding)
     {
