+++ src/materials/EmissiveMaterial.cpp
@@ -4,21 +4,21 @@
 namespace raytracer
 {
 //----------------------------------------------------------------------------------
-EmissiveMaterial::EmissiveMaterial(const Color3d &color, double intensity)
+EmissiveMaterial::EmissiveMaterial(const Color3f &color, float intensity)
     : m_texture(std::make_shared<SolidColorTexture>(color))
     , m_intensity(intensity)
 {
 }
 
 //----------------------------------------------------------------------------------
-EmissiveMaterial::EmissiveMaterial(const std::shared_ptr<Texture> &texture, double intensity)
+EmissiveMaterial::EmissiveMaterial(const std::shared_ptr<Texture> &texture, float intensity)
     : m_texture(texture)
     , m_intensity(intensity)
 {
 }
 
 //----------------------------------------------------------------------------------
-Color3d EmissiveMaterial::emitted(double u, double v, const glm::dvec3 &point) const
+Color3f EmissiveMaterial::emitted(float u, float v, const glm::vec3 &point) const
 {
     return m_intensity * m_texture->value(u, v, point);
 }
